+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 Change Log
 ==========
 
-- 新增 BadRequest、ErrorCode 與 UserError 類別。
+- 新增 Slim\Controller 與 ResponseHelperTrait 類別。
+- 新增 Error\BadRequest、ErrorCode 與 UserError 類別。
 
 ### 0.0.5 (2017-07-03)
 
-- 新增 MonologFluentHandler 類別
+- 新增 Monolog\MonologFluentHandler 類別
 
 ### 0.0.4 (2017-06-21)
 
@@ -13,7 +14,7 @@ Change Log
 
 ### 0.0.3 (2017-06-21)
 
-- 新增 RouteLogger 與 ContainerIds 類別
+- 新增 Slim\Middleware\RouteLogger 與 ContainerIds 類別
 
 ### 0.0.2 (2017-06-19)